requirements.txt
aiosignal==1.3.1
attrs==23.1.0
bidict==0.22.1
boto3==1.26.138
botocore==1.29.138
cap==0.1.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
distlib==0.3.6
filelock==3.12.0
frozenlist==1.3.3
grpcio==1.51.3
idna==3.4
jmespath==1.0.1
jsonschema==4.17.3
msgpack==1.0.5
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
platformdirs==3.5.1
protobuf==4.23.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
ray==2.4.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
tabulate==0.9.0
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.21.0
